+++ src/documentation.spec.ts
@@ -3,7 +3,7 @@ import { config as exampleConfig } from "../example/config";
 import { actions } from "../example/actions";
 import { ActionsFactory } from "./actions-factory";
 import { createSimpleConfig } from "./config";
-import { Documentation } from "./documentation";
+import { Documentation, withExamples } from "./documentation";
 import { z } from "zod";
 import { describe, expect, test, vi } from "vitest";
 import { protocol } from "engine.io";
@@ -269,6 +269,9 @@ describe("Documentation", () => {
               ack: z.tuple([z.any()]).rest(z.any()),
             },
           },
+          examples: {
+            withAck: { payload: [], ack: ["something"] },
+          },
         }),
         actions: [
           factory.build({
@@ -439,6 +442,13 @@ describe("Documentation", () => {
             title: "Testing unsupported types",
           }),
       ).toThrow(`Zod type ${zodType._def.typeName} is unsupported.`);
+    });
+  });
+
+  describe("withExamples()", () => {
+    test("should return the subject if it's a reference or no examples given", () => {
+      expect(withExamples({ $ref: "test" })).toEqual({ $ref: "test" });
+      expect(withExamples({ type: "boolean" })).toEqual({ type: "boolean" });
     });
   });
 });