+++ src/namespace.ts
@@ -1,5 +1,5 @@
 import { z } from "zod";
-import { EmissionMap } from "./emission";
+import { Emission, EmissionMap } from "./emission";
 import { Hooks } from "./hooks";
 
 export const rootNS = "/";
@@ -10,12 +10,21 @@ export const normalizeNS = (name: string): string => {
   return trimmed.startsWith("/") ? trimmed : `/${trimmed}`;
 };
 
+export interface Example<T extends Emission> {
+  payload?: z.output<T["schema"]>;
+  ack?: T["ack"] extends z.AnyZodTuple ? z.input<T["ack"]> : never;
+}
+
 export interface Namespace<E extends EmissionMap, D extends z.SomeZodObject> {
   /**
    * @desc The events that the server can emit
    * @default {}
    * */
   emission: E;
+  /** @desc Examples for the emission per event */
+  examples?: {
+    [K in keyof E]?: Example<E[K]> | Example<E[K]>[];
+  };
   /**
    * @desc Handlers for some events in different contexts
    * @default {}